Tratamentos de error

+++ dia-2/RaizQuadradaSimples.py
@@ -1,6 +1,6 @@
 import math;
 
-numero = float(input("Digite um número: "));
+numero = int(input("Digite um número: "));
 
 raiz = math.sqrt(numero);
